Split up artifacts reducer
Move FormReducer, JSONReducer and JSONParser to
their own file.

Close DCOS-14630

Move JSONParser to JSONReducer of artifacts

Move JSONParser to JSONReducer file to have one
complete Workflow in one file.

Move IpAddress reducer

In the process create tests for the ipAddress
reducer

close DCOS-14695

Move External Volumes reducer

Split up Constraints reducer

Move form Reducer to form reducer

Move Labels reducer

close DCOS-14696

Move Healtchecks Reducers

Move require ports reducer

close DCOS-14729

Move Residency Reducer

close DCOS-14728

Refactor volumes reducer

Remove `FormReducer` as the form uses the specific
typed reducers like external volumes, host volumes
and local volumes.

close DCOS-14727

Move Environment variables reducer

style(ServiceFormReducers): Remove typos

This PR removes typs from the Environment
Variables reducer test cases.

split local volumes reducer

close DCOS-14967
