+++ plugins/services/src/js/reducers/JSONMultiContainerParser.js
@@ -4,17 +4,23 @@ import { JSONParser as container } from "./serviceForm/Container";
 import {
   JSONParser as constraints
 } from "./serviceForm/MultiContainerConstraints";
-import { JSONParser as fetch } from "./serviceForm/Artifacts";
+import { JSONParser as fetch } from "./serviceForm/JSONReducers/Artifacts";
 import {
   JSONParser as environmentVariables
-} from "./serviceForm/MultiContainerEnvironmentVariables";
-import { JSONParser as externalVolumes } from "./serviceForm/ExternalVolumes";
-import { JSONParser as healthChecks } from "./serviceForm/HealthChecks";
-import { JSONParser as labels } from "./serviceForm/Labels";
-import { JSONParser as localVolumes } from "./serviceForm/LocalVolumes";
+} from "./serviceForm/JSONReducers/EnvironmentVariables";
+import {
+  JSONParser as externalVolumes
+} from "./serviceForm/JSONReducers/ExternalVolumes";
+import {
+  JSONParser as healthChecks
+} from "./serviceForm/JSONReducers/HealthChecks";
+import {
+  JSONParser as localVolumes
+} from "./serviceForm/JSONReducers/LocalVolumes";
+import { JSONParser as labels } from "./serviceForm/JSONReducers/Labels";
 import { JSONParser as portDefinitions } from "./serviceForm/PortDefinitions";
 import { JSONParser as portMappings } from "./serviceForm/PortMappings";
-import { JSONParser as residency } from "./serviceForm/Residency";
+import { JSONParser as residency } from "./serviceForm/JSONReducers/Residency";
 import { JSONParser as scaling } from "./serviceForm/MultiContainerScaling";
 import { JSONParser as network } from "./serviceForm/Network";
 import {
