+++ plugins/services/src/js/reducers/JSONMultiContainerReducers.js
@@ -4,13 +4,15 @@ import {
   JSONReducer as constraints
 } from "./serviceForm/MultiContainerConstraints";
 import { JSONReducer as containers } from "./serviceForm/Containers";
-import { JSONReducer as environment } from "./serviceForm/EnvironmentVariables";
-import { JSONReducer as fetch } from "./serviceForm/Artifacts";
+import {
+  JSONReducer as environment
+} from "./serviceForm/JSONReducers/EnvironmentVariables";
+import { JSONReducer as fetch } from "./serviceForm/JSONReducers/Artifacts";
 import { JSONReducer as scaling } from "./serviceForm/MultiContainerScaling";
-import { JSONReducer as labels } from "./serviceForm/Labels";
+import { JSONReducer as labels } from "./serviceForm/JSONReducers/Labels";
 import { JSONReducer as volumes } from "./serviceForm/MultiContainerVolumes";
 import { JSONReducer as networks } from "./serviceForm/MultiContainerNetwork";
-import { JSONReducer as ipAddress } from "./serviceForm/IpAddress";
+import { JSONReducer as ipAddress } from "./serviceForm/JSONReducers/IpAddress";
 
 module.exports = {
   id: simpleReducer("id"),
