+++ plugins/services/src/js/reducers/serviceForm/JSONReducers/Artifacts.js
@@ -33,20 +33,6 @@ function processTransaction(state, { type, path, value }) {
 }
 
 module.exports = {
-  JSONReducer(state, { type, path, value }) {
-    if (path == null) {
-      return state;
-    }
-
-    if (this.fetch == null) {
-      this.fetch = [];
-    }
-
-    this.fetch = processTransaction(this.fetch, { type, path, value });
-
-    return getJson(this.fetch);
-  },
-
   JSONParser(state) {
     if (state.fetch == null) {
       return [];
@@ -62,13 +48,17 @@ module.exports = {
     }, []);
   },
 
-  FormReducer(state = [], { type, path, value }) {
+  JSONReducer(state, { type, path, value }) {
     if (path == null) {
       return state;
     }
 
-    state = processTransaction(state, { type, path, value });
+    if (this.fetch == null) {
+      this.fetch = [];
+    }
 
-    return state;
+    this.fetch = processTransaction(this.fetch, { type, path, value });
+
+    return getJson(this.fetch);
   }
 };
