+++ plugins/services/src/js/reducers/serviceForm/JSONReducers/Constraints.js
@@ -1,10 +1,8 @@
 import { findNestedPropertyInObject } from "#SRC/js/utils/Util";
 import { isEmpty } from "#SRC/js/utils/ValidatorUtil";
-
-import { FormReducer, JSONParser, JSONReducer } from "./common/Constraints";
+import { JSONParser, JSONReducer } from "../common/Constraints";
 
 module.exports = {
-  FormReducer,
   JSONReducer(state, transaction) {
     const constraints = JSONReducer.bind(this)(state, transaction);
