+++ plugins/services/src/js/reducers/serviceForm/JSONReducers/Volumes.js
@@ -170,6 +170,5 @@ function reduceVolumes(state, { type, path, value }) {
 }
 
 module.exports = {
-  JSONReducer: reduceVolumes,
-  FormReducer: reduceVolumes
+  JSONReducer: reduceVolumes
 };
