+++ plugins/services/src/js/reducers/serviceForm/JSONReducers/__tests__/Artifacts-test.js
@@ -1,26 +1,9 @@
-const Artifacts = require("../Artifacts");
+const { SET, ADD_ITEM } = require("#SRC/js/constants/TransactionTypes");
 const Batch = require("#SRC/js/structs/Batch");
 const Transaction = require("#SRC/js/structs/Transaction");
-const { SET, ADD_ITEM } = require("#SRC/js/constants/TransactionTypes");
+const Artifacts = require("../Artifacts");
 
 describe("Artifacts", function() {
-  // #FormReducer behaves the same
-  describe("#JSONReducer", function() {
-    it("emits correct JSON", function() {
-      const batch = new Batch([
-        new Transaction(["fetch"], 0, ADD_ITEM),
-        new Transaction(["fetch", 0, "uri"], "http://mesosphere.io", SET),
-        new Transaction(["fetch"], 1, ADD_ITEM),
-        new Transaction(["fetch", 1, "uri"], "http://mesosphere.com", SET)
-      ]);
-
-      expect(batch.reduce(Artifacts.JSONReducer.bind({}), {})).toEqual([
-        { uri: "http://mesosphere.io" },
-        { uri: "http://mesosphere.com" }
-      ]);
-    });
-  });
-
   describe("#JSONParser", function() {
     it("parses JSON correctly", function() {
       expect(
@@ -35,6 +18,22 @@ describe("Artifacts", function() {
         { type: SET, path: ["fetch", 0, "uri"], value: "http://mesosphere.io" },
         { type: ADD_ITEM, path: ["fetch"], value: 1 },
         { type: SET, path: ["fetch", 1, "uri"], value: "http://mesosphere.com" }
+      ]);
+    });
+  });
+
+  describe("#JSONReducer", function() {
+    it("emits correct JSON", function() {
+      const batch = new Batch([
+        new Transaction(["fetch"], 0, ADD_ITEM),
+        new Transaction(["fetch", 0, "uri"], "http://mesosphere.io", SET),
+        new Transaction(["fetch"], 1, ADD_ITEM),
+        new Transaction(["fetch", 1, "uri"], "http://mesosphere.com", SET)
+      ]);
+
+      expect(batch.reduce(Artifacts.JSONReducer.bind({}), {})).toEqual([
+        { uri: "http://mesosphere.io" },
+        { uri: "http://mesosphere.com" }
       ]);
     });
   });
