+++ plugins/services/src/js/reducers/serviceForm/JSONReducers/__tests__/RequirePorts-test.js
@@ -1,7 +1,7 @@
-const RequirePorts = require("../RequirePorts");
+const { SET } = require("#SRC/js/constants/TransactionTypes");
 const Batch = require("#SRC/js/structs/Batch");
 const Transaction = require("#SRC/js/structs/Transaction");
-const { SET } = require("#SRC/js/constants/TransactionTypes");
+const RequirePorts = require("../RequirePorts");
 
 describe("RequirePorts", function() {
   describe("#JSONReducer", function() {
