+++ plugins/services/src/js/reducers/serviceForm/JSONReducers/__tests__/Residency-test.js
@@ -1,11 +1,11 @@
-const Residency = require("../Residency");
-const Batch = require("#SRC/js/structs/Batch");
-const Transaction = require("#SRC/js/structs/Transaction");
 const {
   ADD_ITEM,
   REMOVE_ITEM,
   SET
 } = require("#SRC/js/constants/TransactionTypes");
+const Batch = require("#SRC/js/structs/Batch");
+const Transaction = require("#SRC/js/structs/Transaction");
+const Residency = require("../Residency");
 
 describe("Residency", function() {
   describe("#JSONReducer", function() {
